add frontend, docker

+++ backend/src/reviews/schemas.py
@@ -7,7 +7,7 @@ class ReviewCreate(BaseModel):
     Schema for creating a new review
     """
     station_id: int
-    rating: int
+    rating: float
     comment: Optional[str] = None
 
 
@@ -17,5 +17,5 @@ class ReviewUpdate(BaseModel):
     Schema for updating an existing review
     """
     station_id: int
-    rating: int
+    rating: float
     comment: Optional[str] = None
